Update to 2.4.6.
This is the first production release of the OpenLDAP 2.4 series,
and it fixes several security issues, as mentioned in vuln.xml
item db449245-870d-11dc-a3ec-001921ab2fa4.

Note that this is generally a no-op for most users because this
version is not (yet) set as default OpenLDAP implementation.

Security:	db449245-870d-11dc-a3ec-001921ab2fa4
Approved by:	portmgr (linimon)