- Change LUA_PREFIX default value from LOCALBASE to PREFIX.
Actually the Lua prefix should be recorded on each port and read from there
(as with "*-config --prefix" scripts), instead of this assumption. But
supporting PREFIX is a requirement, while supporting different custom
combinations of a general PREFIX for dependencies and a different PREFIX for a
specific port is not. Also the latter could be achieved by changing LOCALBASE
too, or directly modifying LUA_PREFIX in this case.

PR:		ports/120277
Submitted by:	Hirohisa Yamaguchi <umq@ueo.co.jp>