A user may now build a list of rules that will be used to determine if unsafe images (that are linked to remote sites) will be shown in HTML messages. If a message matches any of the rules and contains images that would normally be initially hidden, then they are now shown by default.
The user may choose to always show unsafe images, for all message. This is
obviously not recommended by the core SquirrelMail Project Team - or they
wouldn't have built this functionality to begin with ( See the following:
http://www.squirrelmail.org/wiki/UnsafeImages ).

A new section is added to the options page titled, 'Unsafe Image Rules'. Within
this page the user may define a number of rules to determine when messages are
from a trusted source.

These options are very similar to the core message filters plugin. A message
field (To, From, CC, Subject) can be matched either against a regular
expression, or simply searched to see if the given string is within the field.
If a match is found then unsafe images are always shown for this source.

WWW: http://www.squirrelmail.org/plugin_view.php?id=98