Fix mouse probing issues with HAL support:
- Do not probe /dev/psm0 as SysMouse if HAL support is enabled.
- Add /dev/ums0 as SysMouse if HAL support is disabled.
- Do not force sysmouse(4) mode unless device is /dev/sysmouse.  When
moused(8) is not running, current mode level is untouched.
- Correcly check if moused is actually running.  Checking pid file is
not enough because it may be /var/run/moused.<device>.pid when it is
invoked by devd(8) for USB mice.
- Do not attach device if it is opened by moused.
- Correctly add device path for HAL.

This patch is tested with many possible combinations of moused(8), HAL,
PS/2 mouse, USB mouse, and xorg.conf.

Tested by:	cokane, jkim, marcus,
		Matthew Seaman (m dot seaman at infracaninophile dot co dot uk)