New port: p5-Net-SMTP-SSL
Implements the same API as Net::SMTP, but uses IO::Socket::SSL for its
network operations. Due to the nature of Net::SMTP's new method, it is
not overridden to make use of a default port for the SMTPS service.
Perhaps future versions will be smart like that. Port 465 is usually
what you want, and it's not a pain to specify that.

WWW: http://search.cpan.org/dist/Net-SMTP-SSL/